........
  r8290 | sinbad | 2009-01-27 15:45:47 +0000 (Tue, 27 Jan 2009) | 1 line

  Patch 2533453, final part only - safety check when using mCurrentContext
........
  r8291 | sinbad | 2009-01-27 16:35:21 +0000 (Tue, 27 Jan 2009) | 1 line

  Patch 2512292: standardise use of NOMINMAX and WIN32_LEAN_AND_MEAN when including windows.h
........
  r8293 | sinbad | 2009-01-27 17:11:26 +0000 (Tue, 27 Jan 2009) | 1 line

  Fix small potential memory leak when explicitly constructing a SharedPtr with a null pointer (thanks bishopnator)
........
  r8295 | sinbad | 2009-01-29 11:40:09 +0000 (Thu, 29 Jan 2009) | 1 line

  Fix a shadow map indexing bug when using pass iteration for a specific light type and multiple light types all with shadows
........
  r8296 | sinbad | 2009-01-29 12:40:11 +0000 (Thu, 29 Jan 2009) | 1 line

  Remove flushUpTo() call in CompositorChain::postTargetOperation; this happens after the begin/end of the scene and actually gets ignored anyway, D3D9 would bitch about this in debug mode
........
  r8299 | sinbad | 2009-01-30 11:16:37 +0000 (Fri, 30 Jan 2009) | 1 line

  Return float from AutoParamDataSource::getLightCastsShadows to match declared type
........
  r8306 | sinbad | 2009-01-31 18:30:24 +0000 (Sat, 31 Jan 2009) | 1 line

  Patch 2525593: make methods of TerrainZone virtual
........
  r8313 | sinbad | 2009-02-09 14:44:11 +0000 (Mon, 09 Feb 2009) | 1 line

  Fix off-by-one error in StaticGeometry::GeometryBucket::assign. Should test the vertex counts against the mMaxVertexIndex + 1
........
  r8315 | sinbad | 2009-02-09 16:11:48 +0000 (Mon, 09 Feb 2009) | 1 line

  Fix for VC7.1 threaded build, VC7.1 doesn't have intrin.h
........
  r8316 | sinbad | 2009-02-12 16:22:08 +0000 (Thu, 12 Feb 2009) | 1 line

  Fix crash with Maya exporter (thanks dpakatheman)
........
  r8317 | Lf3T-Hn4D | 2009-02-12 17:21:47 +0000 (Thu, 12 Feb 2009) | 1 line

  - Fix exporter failing when exporting armature animations with no actions being selected on the armature prior to export.
........
  r8318 | Lf3T-Hn4D | 2009-02-12 18:04:13 +0000 (Thu, 12 Feb 2009) | 1 line

  - reset action to None after exporting if it was not set before export.
........

+++ OgreMain/include/OgreAutoParamDataSource.h
@@ -169,7 +169,7 @@ namespace Ogre {
         /** Get the light which is 'index'th closest to the current object */        
 		virtual float getLightNumber(size_t index) const;
 		virtual float getLightCount() const;
-		virtual int getLightCastsShadows(size_t index) const;
+		virtual float getLightCastsShadows(size_t index) const;
 		virtual const ColourValue& getLightDiffuseColour(size_t index) const;
 		virtual const ColourValue& getLightSpecularColour(size_t index) const;
 		virtual const ColourValue getLightDiffuseColourWithPower(size_t index) const;
